+++ tests/storage/test_media_operations.py
@@ -159,8 +159,9 @@ VIOS_DOC_WITH_OPTICAL_MAPPINGS = f"""<?xml version="1.0" encoding="UTF-8" standa
   <id>urn:uuid:{VIOS_UUID}</id>
   <content type="application/vnd.ibm.powervm.uom+xml">
     <VirtualIOServer xmlns="http://www.ibm.com/xmlns/systems/power/firmware/uom/mc/2012_10/">
-      <UUID>{VIOS_UUID}</UUID>
+      <Metadata><Atom><AtomID>{VIOS_UUID}</AtomID></Atom></Metadata>
       <PartitionName>vios1</PartitionName>
+      <PartitionUUID kb="ROO">{VIOS_UUID}</PartitionUUID>
       <AssociatedManagedSystem rel="related"
         href="https://hmc/rest/api/uom/ManagedSystem/{SYSTEM_UUID}"/>
       <VirtualSCSIMappings>